\name{robflreg-package}
\alias{robflreg-package}
\alias{robflreg}
\docType{package}
\title{
Robust function-on-function regression
}
\description{
This package presents robust methods for analyzing functional linear regression.
}
\author{
Ufuk Beyaztas and Han Lin Shang

Maintainer: Ufuk Beyaztas <ufukbeyaztas@gmail.com>
}
\references{
U. Beyaztas, A. Mandal and H. L. Shang (2026+) Enhancing spatial functional linear regression with robust dimension reduction methods, \emph{Journal of Multivariate Analysis}, \bold{in press}.

U. Beyaztas, H. L. Shang and S. Saricam (2025) Penalized function-on-function linear quantile regression, \emph{Computational Statistics}, \bold{40}, 301-329.

B. Akturk, U. Beyaztas, H. L. Shang, A. Mandal (2025) Robust functional logistic regression, \emph{Advances in Data Analysis and Classification}, \bold{19}, 121-145.

U. Beyaztas, H. L. Shang and A. Mandal (2025) Robust function-on-function interaction regression, \emph{Statistical Modelling: An International Journal}, \bold{25}(3), 195-215.

M. Mutis, U. Beyaztas, F. Karaman and H. L. Shang (2025) On function-on-function linear quantile regression, \emph{Journal of Applied Statistics}, \bold{52}(4), 814-840.

M. Mutis, U. Beyaztas, G. G. Simsek, H. L. Shang and Z. M. Yaseen (2024) Development of functional quantile autoregressive model for river flow curve forecasting, \emph{Earth and Space Science}, \bold{11}, e2024EA003564.

S. Gurer, H. L. Shang, A. Mandal and U. Beyaztas (2024) Locally sparse and robust partial least squares in scalar-on-function regression, \emph{Statistics and Computing}, \bold{34}, article number 150.

U. Beyaztas, M. Tez and H. L. Shang (2024) Robust scalar-on-function partial quantile regression, \emph{Journal of Applied Statistics}, \bold{51}(7), 1359-1377.

U. Beyaztas and H. L. Shang (2023) Robust functional linear regression models, \emph{The R Journal}, \bold{15}(1), 212-233.

M. Mutis, U. Beyaztas, G. G. Simsek and H. L. Shang (2023) A robust scalar-on-function logistic regression for classification, \emph{Communications in Statistics - Theory and Methods}, \bold{52}(23), 8538-8554.

U. Beyaztas and H. L. Shang (2022) Robust bootstrap prediction intervals for univariate and multivariate autoregressive time series models, \emph{Journal of Applied Statistics}, \bold{49}(5), 1179-1202.

U. Beyaztas, H. L. Shang and A. G. Abdel-Salam (2022) Functional linear model for interval-valued data, \emph{Communications in Statistics - Simulation and Computation}, \bold{51}(7), 3513-3532.

U. Beyaztas and H. L. Shang (2022) Machine-learning-based functional time series forecasting: Application to age-specific mortality rates, \emph{Forecasting}, \bold{4}, 394-408.

S. Saricam, U. Beyaztas, B. Asikgil and H. L. Shang (2022) On partial least-squares estimation in scalar-on-function regression models, \emph{Journal of Chemometrics}, \bold{36}(12), e3452.

U. Beyaztas and H. L. Shang (2022) A comparison of parameter estimation in function-on-function regression, \emph{Communications in Statistics - Simulation and Computation}, \bold{51}(8), 4607-4637.

U. Beyaztas and H. L. Shang (2022) A robust functional partial least squares for scalar-on-multiple-function regression, \emph{Journal of Chemometrics}, \bold{36}(4), e3394.

U. Beyaztas, H. L. Shang and A. Alin (2022) Function-on-function partial quantile regression, \emph{Journal of Agricultural, Biological, and Environmental Statistics}, \bold{27}(1), 149-174.

U. Beyaztas, H. L. Shang and Z. M. Yaseen (2021) A functional autoregressive model based on exogenous hydrometeorological variables for river flow prediction, \emph{Journal of Hydrology}, \bold{598}, 126380.

U. Beyaztas and H. L. Shang (2021) A partial least squares approach for function-on-function interaction regression, \emph{Computational Statistics}, \bold{36}(2), 911-939.

U. Beyaztas and H. L. Shang (2021) A robust partial least squares approach for function-on-function regression, \emph{Brazilian Journal of Probability and Statistics}, \bold{36}(2), 199-219.

U. Beyaztas and H. L. Shang (2021) Function-on-function linear quantile regression, \emph{Mathematical Modelling and Analysis}, \bold{27}(2), 322-341.

U. Beyaztas and H. L. Shang (2020) On function-on-function regression: partial least squares approach, \emph{Environmental and Ecological Statistics}, \bold{27}(1), 95-114.

U. Beyaztas and H. L. Shang (2019) Forecasting functional time series using weighted likelihood methodology, \emph{Journal of Statistical Computation and Simulation}, \bold{89}(16), 3046-3060.
}
\keyword{package}
